More refactoring and addition of JS functionality
I'm coming up on the initial 1.0 release here and about ready for heavy testing, then writing some tests and documentation, and setting up a github page and then we'll call it 1.0 and then all that'll be left to do is mapping the rest of Selenium's features like cookies and such. All the core functionality is here at this point and the boilerplate and technical debt of selenium is all cleanly abstracted away from the end user in an easy to use manner that's easily modifiable.